deps: update V8 to 14.7.66

+++ deps/v8/src/objects/deoptimization-data.cc
@@ -153,16 +153,19 @@ size_t DeoptimizationLiteral::Read(base::Vector<const uint8_t> buffer,
   UNREACHABLE();
 }
 
+// TODO(375937549): Convert deopt_entry_count to uint32_t.
 Handle<DeoptimizationData> DeoptimizationData::New(Isolate* isolate,
                                                    int deopt_entry_count) {
   return TrustedCast<DeoptimizationData>(
-      isolate->factory()->NewProtectedFixedArray(LengthFor(deopt_entry_count)));
+      isolate->factory()->NewProtectedFixedArray(
+          static_cast<uint32_t>(LengthFor(deopt_entry_count))));
 }
 
 Handle<DeoptimizationData> DeoptimizationData::New(LocalIsolate* isolate,
                                                    int deopt_entry_count) {
   return TrustedCast<DeoptimizationData>(
-      isolate->factory()->NewProtectedFixedArray(LengthFor(deopt_entry_count)));
+      isolate->factory()->NewProtectedFixedArray(
+          static_cast<uint32_t>(LengthFor(deopt_entry_count))));
 }
 
 Handle<DeoptimizationData> DeoptimizationData::Empty(Isolate* isolate) {
